+++ packages/widget/src/pages/ErrorWarningPage/ExpectedErrorPage/ExpectedErrorPageAuthFailed.tsx
@@ -2,10 +2,10 @@ import { ErrorWarningPageContent } from "@/pages/ErrorWarningPage/ErrorWarningPa
 import { MainButton } from "@/components/MainButton";
 import { ICONS } from "@/icons";
 import { useTheme } from "styled-components";
-import { SwapPageHeader } from "../../SwapPage/SwapPageHeader";
 import { useSetAtom } from "jotai";
 import { errorWarningAtom } from "@/state/errorWarning";
 import { track } from "@amplitude/analytics-browser";
+import { PageHeader } from "@/components/PageHeader";
 
 export type ExpectedErrorPageAuthFailedProps = {
   onClickBack: () => void;
@@ -23,7 +23,7 @@ export const ExpectedErrorPageAuthFailed = ({ onClickBack }: ExpectedErrorPageAu
 
   return (
     <>
-      <SwapPageHeader
+      <PageHeader
         leftButton={{
           label: "Back",
           icon: ICONS.thinArrow,